refactor: Write API function
