+++ src/apis/User/User.ts
@@ -1,0 +1,7 @@
+import requester from "../requester";
+
+export default {
+  getMe() {
+    return requester("/users/me/");
+  },
+};
